django/requirements.txt
Django==3.1.0
djangorestframework
django-filter
django-cors-headers
psycopg2
pylint
pylint-django